[PROD-3970] break down testing into group
