+++ test/tools/driverLicenseTestScenarios.ts
@@ -1,0 +1,315 @@
+import fs from 'fs';
+import { faker } from '@faker-js/faker';
+import jsf from 'json-schema-faker';
+import { yyyymmdd } from './utils';
+
+const dlSchema = JSON.parse(
+  fs.readFileSync('./schemas/json/DriversLicenceCredential.json', 'utf-8')
+);
+
+function generateDriversLicenceCredential(callback?: (data: any) => void): any {
+  const data = jsf.generate(dlSchema) as any;
+  const did = `did:receptor:redbelly:${faker.helpers.arrayElement([
+    'testnet',
+    'mainnet',
+  ])}:${faker.string.alphanumeric(42)}`;
+
+  data['@context'] =
+    'https://raw.githubusercontent.com/redbellynetwork/receptor-schema/refs/heads/main/schemas/json-ld/DriversLicenceCredential.jsonld';
+  data.id = faker.string.uuid();
+  data.type = ['VerifiableCredential', 'DriversLicenceCredential'];
+  data.issuanceDate = faker.date.past().toISOString();
+  data.expirationDate = faker.date.future().toISOString();
+  data.issuer = { id: faker.internet.url() };
+  data.credentialSchema = {
+    id: 'https://raw.githubusercontent.com/redbellynetwork/receptor-schema/refs/heads/main/schemas/json/DriversLicenceCredential.json',
+    type: 'JsonSchemaValidator2018',
+  };
+  data.credentialStatus = {
+    id: faker.internet.url(),
+    type: 'CredentialStatusList2021',
+  };
+  data.subjectPosition = faker.helpers.arrayElement(['none', 'index', 'value']);
+  data.merklizationRootPosition = faker.helpers.arrayElement([
+    'none',
+    'index',
+    'value',
+  ]);
+  data.revNonce = faker.number.int();
+  data.version = faker.number.int();
+  data.updatable = faker.datatype.boolean();
+
+  data.credentialSubject = {
+    id: did,
+    name: faker.person.fullName(),
+    licenceNumber: faker.string.uuid(),
+    country: faker.location.state(),
+    birthDate: yyyymmdd(faker.date.recent({ days: 10 })),
+    expiryDate: yyyymmdd(faker.date.soon({ days: 5 })),
+    publicAddress: faker.finance.ethereumAddress(),
+  };
+
+  if (callback) {
+    callback(data);
+  }
+
+  return data;
+}
+
+
+export const dLTestScenarios = [
+  {
+    name: 'Valid Drivers Licence',
+    data: generateDriversLicenceCredential(),
+    expectedValid: true,
+  },
+  {
+    name: 'Missing Required Field',
+    data: generateDriversLicenceCredential((data) => {
+      delete data.credentialSubject.licenceNumber;
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Wrong Data Type',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.birthDate = 'not_a_date';
+    }),
+    expectedValid: false,
+  },
+  // {
+  //   name: 'Malformed Context: Invalid string',
+  //   data: generateDriversLicenceCredential((data) => {
+  //     data['@context'] = 'invalid_context';
+  //   }),
+  //   expectedValid: false,
+  // },
+  // {
+  //   name: 'Malformed Context: Empty array',
+  //   data: generateDriversLicenceCredential((data) => {
+  //     data['@context'] = [];
+  //   }),
+  //   expectedValid: false,
+  // },
+  {
+    name: 'Malformed Context: undefined',
+    data: generateDriversLicenceCredential((data) => {
+      data['@context'] = undefined;
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid URL in id',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.id = 'not_a_url';
+    }),
+    expectedValid: false,
+  },
+  // {
+  //   name: 'Extreme Values',
+  //   data: generateDriversLicenceCredential((data) => {
+  //     data.credentialSubject.licenceNumber = 'X'.repeat(1000);
+  //   }),
+  //   expectedValid: false,
+  // },
+  {
+    name: 'Null Values',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = null;
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Incorrect Enum Value',
+    data: generateDriversLicenceCredential((data) => {
+      data.subjectPosition = 'maybe';
+    }),
+    expectedValid: false,
+  },
+  // {
+  //   name: 'Injection Attack (XSS)',
+  //   data: generateDriversLicenceCredential((data) => {
+  //     data.credentialSubject.name = "<script>alert('XSS')</script>";
+  //   }),
+  //   expectedValid: false,
+  // },
+  // {
+  //   name: 'Invalid @type',
+  //   data: generateDriversLicenceCredential((data) => {
+  //     data.credentialSubject.type = 'SomeOtherType';
+  //   }),
+  //   expectedValid: false,
+  // },
+  {
+    name: 'Missing id',
+    data: generateDriversLicenceCredential((data) => {
+      delete data.credentialSubject.id;
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Leading/Trailing Spaces',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = ' NSW ';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.id: invalid string',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.id = 'did:In-valid string:redbelly';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.id: empty string',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.id = '';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.name: empty',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.name = '';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.licenceNumber: leading spaces',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.licenceNumber = ' QWER12345';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.licenceNumber: trailing spaces',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.licenceNumber = 'QWER12345 ';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.licenceNumber: empty string',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.licenceNumber = '';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.licenceNumber: spaces in between',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.licenceNumber = 'QWER 12345';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.country: spaces in between',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = "Martha's vineyard";
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.country: empty string',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = '';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.country: trailing space',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = 'CA ';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.country: leading space',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = ' California';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.country: hyphen not allowed',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = 'New-York';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid credentialSubject.country: numbers not allowed',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.country = '1234';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid birthDate: Too short',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.birthDate = 10101;
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid birthDate: Too long',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.birthDate = 199901011;
+    }),
+    expectedValid: false,
+  },
+  // {
+  //   name: 'Invalid birthDate: Invalid date (Feb 30)',
+  //   data: generateDriversLicenceCredential((data) => {
+  //     data.credentialSubject.birthDate = 20240230;
+  //   }),
+  //   expectedValid: false,
+  // },
+  // {
+  //   name: 'Invalid birthDate: MMDDYYYY instead of YYYYMMDD',
+  //   data: generateDriversLicenceCredential((data) => {
+  //     data.credentialSubject.birthDate = 12022024;
+  //   }),
+  //   expectedValid: false,
+  // },
+  {
+    name: 'Invalid publicAddress: Too short',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.publicAddress = '0x123';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid publicAddress: Contains non-hex characters',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.publicAddress =
+        '0xGHIJKL7890abcdef1234567890abcdef12345678';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid publicAddress: Missing 0x prefix',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.publicAddress =
+        '1234567890abcdef1234567890abcdef12345678';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Invalid publicAddress: Uppercase 0X prefix',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.publicAddress =
+        '0XABCDEF1234567890ABCDEF1234567890ABCDEF12';
+    }),
+    expectedValid: false,
+  },
+  {
+    name: 'Valid publicAddress: Mixed case (allowed in Ethereum)',
+    data: generateDriversLicenceCredential((data) => {
+      data.credentialSubject.publicAddress =
+        '0xAbCdEf1234567890ABCDEF1234567890abcdef12';
+    }),
+    expectedValid: true,
+  },
+];
